Исправления после ревью: изменен класс RentalDetailsPage (разбит на более детальные методы + написан общий класс заполнения формы на основе эти методов), убрала использование метода System.out.println где он был излишним, вынесла в BaseTest методы с аннотациями @Before и @After, дополнила файл readme, убрала кастомизированный путь до файла с драйвером браузера
