+++ README.md
@@ -1,3 +1,20 @@
 # YPSprint_4
-Спринт 4ый. 
-Тестируем что-то
+
+Это финальный проект 4го спринта курса "Яндекс практикум"
+Тестируем веб сайт аренды электросамокатов. 
+Тесты написаны с использованием Selenium для взаимодействия с веб-браузером и JUnit для организации и выполнения тестов.
+
+
+Проект использует технологии:
+
+- Java 11
+- JUnit 4.13.2
+- Selenium 3.141.59
+- Maven
+
+
+Структура проекта:
+
+src/main/java/pageobject - пакет POM, содержит классы для описания тестовых методов и локаторов
+src/test/java - тесты
+pom.xml - файл конфигурации Maven с зависимостями и настройками проекта
