+++ pom.xml
@@ -22,7 +22,6 @@
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
             <version>4.13.2</version>
-            <scope>test</scope>
         </dependency>
 
         <!-- Зависимости для Selenium -->
